requirements.txt
python-dotenv>=0.19.0
opencv-python>=4.5.0
requests>=2.26.0
pymongo>=3.12.0
psutil>=5.8.0
python-telegram-bot>=13.7
pytest>=6.2.5
pillow>=8.3.2
numpy>=1.21.0 